Ensure that a selection is always random
As it was, select() took the first elements as the provisional selection
in iteration order. That would not make for a *random* selection in my
estimation.

Instead of requiring a 'sample' or 'shuffle' function as argument,
I decided to shuffle by hand using randbelow().

First, the documentation for random.shuffle() is not reassuring.
While the code uses Random.randbelow() (and yields a warning on recent
versions of Python when passing a 'random' argument), the docstring
does not reflect that and still references random.random as the "default".
So, what default shuffle function should I use?

Second, the other argument would be redundant, since shuffling is naturally
implemented with repeated randbelow() calls. Requiring a full-blown Random
object is not an option, either.

Third, the code is very simple.