Support UTF-8 (not just ASCII)
The changes alter the API, which now speaks characters instead of
bytes, so version bumps to 0.4.0.

An 'encoding' optional argument is added to dictionary().

Passphrase now uses lists of characters for representing words
internally instead of bytearrays. They are converted to strings
and back to lists when necessary: which is ugly but works.

Methods of Passphrase change accordingly:
  - parameters which were 'bytes' now must be strings;
  - replace() accepts functions of strings, str -> str;
  - translate() conforms to str.translate(), which misses the delete
    parameter but takes a different form of table.

All the surrounding functionality deals in characters:
  - The ord_range() function is replaced by char_range().
  - parse_charset() returns sets of characters.

(Fin.)